requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-browser-reload==1.15.0
django-constance==4.1.2
django-debug-toolbar==4.4.6
django-storages==1.14.4
django-unfold==0.40.0
Faker==29.0.0
idna==3.10
isodate==0.7.2
kombu==5.4.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
python-dateutil==2.9.0.post0
redis==5.1.1
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
watchdog==5.0.3
wcwidth==0.2.13
